feat: shop category에 따른 detail pages 통일
