+++ src/utils/register/scrollUp.ts
@@ -1,4 +1,4 @@
-// 스크롤 제일 위로 (추후 다른 페이지에도 쓰일수 있을거 같아서 일단 util에 생성)
-export const scrollToTop = () => {
-  window.scrollTo(0, 0);
+// 스크롤 제일 위로
+export const scrollToTop = (height?: number) => {
+  height? window.scrollTo(0, height) : window.scrollTo(0, 0);
 };
